+++ apps/web/src/widgets/header/index.ts
@@ -1,2 +1,2 @@
-export { HomeHeader } from "./ui/Tmp_homeHeader";
-export { PostHeader } from "./ui/Tmp_postHeader";
+export { HomeHeader } from "./ui/HomeHeader";
+export { PostHeader } from "./ui/PostHeader";
